+++ channels/boteco-channel-telegram/src/main/java/tools/devnull/boteco/channel/telegram/TelegramIncomeProcessor.java
@@ -27,10 +27,7 @@ package tools.devnull.boteco.channel.telegram;
 import org.apache.camel.Exchange;
 import org.apache.camel.Processor;
 import tools.devnull.boteco.Channel;
-import tools.devnull.boteco.Destination;
 import tools.devnull.boteco.message.MessageDispatcher;
-import tools.devnull.boteco.user.User;
-import tools.devnull.boteco.user.UserManager;
 
 /**
  * A processor that deals with income messages from Telegram.
@@ -40,7 +37,6 @@ public class TelegramIncomeProcessor implements Processor {
   private final Channel channel;
   private final TelegramOffsetManager offsetManager;
   private final MessageDispatcher dispatcher;
-  private final UserManager userManager;
 
   /**
    * Creates a new processor based on the given parameters
@@ -48,25 +44,20 @@ public class TelegramIncomeProcessor implements Processor {
    * @param channel the channel implementation
    * @param offsetManager   a component to manager the current offset in poll operations
    * @param dispatcher      a component to dispatch messages to be processed
-   * @param userManager     a user manager to fetch user information
    */
   public TelegramIncomeProcessor(Channel channel,
                                  TelegramOffsetManager offsetManager,
-                                 MessageDispatcher dispatcher,
-                                 UserManager userManager) {
+                                 MessageDispatcher dispatcher) {
     this.channel = channel;
     this.offsetManager = offsetManager;
     this.dispatcher = dispatcher;
-    this.userManager = userManager;
   }
 
   @Override
   public void process(Exchange exchange) throws Exception {
     offsetManager.process(exchange.getIn().getBody(TelegramPolling.class),
         pooling -> {
-          String senderId = pooling.getMessage().getChat().getId().toString();
-          User user = this.userManager.find(Destination.channel(TelegramChannel.ID).to(senderId));
-          dispatcher.dispatch(new TelegramIncomeMessage(channel, pooling.getMessage(), user));
+          dispatcher.dispatch(new TelegramIncomeMessage(channel, pooling.getMessage()));
         });
   }