+++ plugins/boteco-plugin-redhat/src/main/java/tools/devnull/boteco/plugins/redhat/RedHatPlugin.java
@@ -25,14 +25,12 @@
 package tools.devnull.boteco.plugins.redhat;
 
 import tools.devnull.boteco.plugin.Listener;
-import tools.devnull.boteco.plugin.Notification;
 import tools.devnull.boteco.plugin.Plugin;
 
 import java.util.Collections;
 import java.util.List;
 
 import static tools.devnull.boteco.plugin.Listener.listenTo;
-import static tools.devnull.boteco.plugin.Notification.notifies;
 
 public class RedHatPlugin implements Plugin {
 
@@ -51,13 +49,6 @@ public class RedHatPlugin implements Plugin {
     return Collections.singletonList(
         listenTo("Occurrences of links from access.redhat.com articles and solutions")
             .respondWith("Detailed information about the article/solution")
-    );
-  }
-
-  @Override
-  public List<Notification> notifications() {
-    return Collections.singletonList(
-        notifies("status.redhat").about("Incidents from status.redhat.com")
     );
   }